remove unsafe_cast_to_jsonrpsee_client

+++ algorithms/beefy/prover/src/helpers.rs
@@ -82,13 +82,13 @@ pub async fn fetch_timestamp_extrinsic_with_proof<T: Config>(
 				<TrieDBMutBuilder<sp_trie::LayoutV0<BlakeTwo256>>>::new(&mut db, &mut root).build();
 
 			for (i, ext) in extrinsics.into_iter().enumerate() {
-				let key = codec::Compact(i as u32).encode();
+				let key = codec::Compact(i as u64).encode();
 				trie.insert(&key, &ext)?;
 			}
 			*trie.root()
 		};
 
-		let key = codec::Compact::<u32>(0u32).encode();
+		let key = codec::Compact::<u64>(0u64).encode();
 		let extrinsic_proof =
 			generate_trie_proof::<sp_trie::LayoutV0<BlakeTwo256>, _, _, _>(&db, root, vec![&key])?;
